content/canvas/test/test_2d.composite.globalAlpha.invalid.html

<!DOCTYPE HTML>
<title>Canvas test: 2d.composite.globalAlpha.invalid</title>
<script src="/MochiKit/packed.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

ctx.globalAlpha = 0.5;
var a = ctx.globalAlpha; // might not be exactly 0.5, if it is rounded/quantised, so remember for future comparisons
var _thrown = undefined; try {
  ctx.globalAlpha = Infinity;
} catch (e) { _thrown = e }; todo(_thrown && _thrown.code == DOMException.NOT_SUPPORTED_ERR, "should throw NOT_SUPPORTED_ERR");
ok(ctx.globalAlpha == a, "ctx.globalAlpha == a");
var _thrown = undefined; try {
  ctx.globalAlpha = -Infinity;
} catch (e) { _thrown = e }; todo(_thrown && _thrown.code == DOMException.NOT_SUPPORTED_ERR, "should throw NOT_SUPPORTED_ERR");
ok(ctx.globalAlpha == a, "ctx.globalAlpha == a");
var _thrown = undefined; try {
  ctx.globalAlpha = NaN;
} catch (e) { _thrown = e }; todo(_thrown && _thrown.code == DOMException.NOT_SUPPORTED_ERR, "should throw NOT_SUPPORTED_ERR");
ok(ctx.globalAlpha == a, "ctx.globalAlpha == a");

SimpleTest.finish();

});
</script>
